Sweet Alert Added

+++ src/app/section/section.component.ts
@@ -3,6 +3,7 @@ import { FilterPipe } from '../filter.pipe';
 import { ColorBorderDirective } from '../directive/color-border.directive';
 import { CommonModule } from '@angular/common';
 import { CourseService } from '../course.service'
+import  Swal  from 'sweetalert2';
 
 @Component({
   selector: 'app-section',
@@ -25,9 +26,34 @@ export class SectionComponent implements OnInit, OnChanges {
     this.posts = _service.getList();
    }
    deleteCourseById(id:number){    
-      this.cursos =  this._service.deleteCourseById(this.posts , id); 
-      this.posts = [];
-      this.posts = this.cursos;  
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You will not be able to recover this imaginary file!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, keep it'
+    }).then((result) => {
+      if (result.value) {
+        this.cursos =  this._service.deleteCourseById(this.posts , id); 
+        this.posts = [];
+        this.posts = this.cursos;  
+        Swal.fire(
+          'Deleted!',
+          'Your imaginary file has been deleted.',
+          'success'
+        )
+      // For more information about handling dismissals please visit
+      // https://sweetalert2.github.io/#handling-dismissals
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.fire(
+          'Cancelled',
+          'Your imaginary file is safe :)',
+          'error'
+        )
+      }
+    })
+     
     }
    ngOnChanges(changes:SimpleChanges){
